% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_cell_time.R
\name{align_cell_time}
\alias{align_cell_time}
\title{Align cell time}
\usage{
align_cell_time(trajectory_expression)
}
\arguments{
\item{trajectory_expression}{a list of length \eqn{N},where \eqn{N} is
the number of cells in one time point(the number of cells need to be equal
among different time points). The \eqn{i}-th element in the list is a matrix with
\eqn{d} rows and \eqn{t} columns, giving the gene expression of \eqn{d} genes in \eqn{t} time
points of the \eqn{i}-th cell lineages trajectory. This list
could be generated by function \code{\link{trajectory_expression}}.}
}
\value{
a matrix of aligned cell time.
}
\description{
\code{align_cell_time()} normalizes the time line since different
cells need different time in the differentiate process.
}
