% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_name_traj.R
\name{cell_name_traj}
\alias{cell_name_traj}
\title{Optimal transport across different time points}
\usage{
cell_name_traj(data)
}
\arguments{
\item{data}{a dataframe with \eqn{N*t} rows and \eqn{d+2} columns, where \eqn{N} is
the number of cells in one time point(the number of cells need to be equal
among different time points),
\eqn{t} is the number of time points, \eqn{d} is the number of genes. The first column
is the cell names and the last column is the time points which contains continuous
integers beginning from 1. The column names of the first and the last columns need to be set as
'index' and 'time', respectively.
This dataframe could be regraded as a result of \code{rbind()} of cell-gene
dataset ordered by \emph{increasing} time points.}
}
\value{
a dataframe with \eqn{N} rows and \eqn{t} columns, giving \eqn{N} unique
trajectories for individual cells, each trajectory has length \eqn{t}.
}
\description{
\code{cell_name_traj()} accomplishes a cell lineage tracing method, which aligns
two individual cells between any two adjacent time points via the optimal
transport technique(Earth mover's distance is used in this function), and returns
unique trajectories for individual cells.

Note that users need to set up a conda environment with package "pot" installed
before using this function.
}
