% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community_connect.R
\name{community_connect}
\alias{community_connect}
\title{Connecting Community between predicted time points}
\usage{
community_connect(group_label)
}
\arguments{
\item{group_label}{a list of length \code{npred}, where \code{npred} is the number of
predicted time points. Each element in the list is a dataframe containing
genes' community belongings in one predicted time point. This list
could be generated by function \code{\link{gene_community_detection}}.}
}
\value{
a list of length \code{npred-1}. The \code{i}-th element in the list is a
dataframe which the first column is community id in the \code{i}-th predicted
time point and the second column is the corresponding community in the \code{(i+1)}-th
predicted time point. Community in the \code{(i+1)}-th time point which shares the
most common genes with community in the \eqn{i}-th time point is regraded as
corresponding community.
}
\description{
\code{community_connect()} constructs connections of communities
between all predicted time points.
}
