% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_linear_relationship.R
\name{dynamic_linear_relationship}
\alias{dynamic_linear_relationship}
\title{Dynamic linear relationship between genes}
\usage{
dynamic_linear_relationship(data, gene_expression, alpha = 0.3, npred = 101)
}
\arguments{
\item{data}{a dataframe with \eqn{N*t} rows and \eqn{d+2} columns, where \eqn{N} is
the number of cells in one time point(the number of cells need to be equal
among different time points),
\eqn{t} is the number of time points, \eqn{d} is the number of genes. The first column
is the cell names and the last column is the time points which contains continuous
integers beginning from 1. The column names of the first and the last columns need to be set as
'index' and 'time', respectively.
This dataframe could be regraded as a result of \code{rbind()} of cell-gene
dataset ordered by \emph{increasing} time points.}

\item{gene_expression}{a list of length \eqn{d}, where \eqn{d} is the number of genes. The elements
in the list representing the individual gene expression in \eqn{N} cell
lineages trajectories, each trajectory has \eqn{t} time points. This list
could be generated by function \code{\link{gene_expression}}.}

\item{alpha}{a parameter used in SSANOVA.Larger absolute values yield smoother fits.
See \code{\link[gss]{ssanova}} for details.}

\item{npred}{the number of predicted time points equally spaced
between 1 and \eqn{t}.}
}
\value{
a dataframe containing strength of pairwise-gene linear relationship in
each predicted time points.
}
\description{
\code{dynamic_linear_relationship()} calculates dynamic strength of
pairwise-gene linear relationship, using functional concurrent linear model and
smoothing spline ANOVA model.
}
