% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_community_detection.R
\name{gene_community_detection}
\alias{gene_community_detection}
\title{Gene community detection}
\usage{
gene_community_detection(network)
}
\arguments{
\item{network}{a list of length \code{npred}, where \code{npred} is the number of
predicted time points. Each element in the list is an object of class 'network'.
This list could be generated by function \code{\link{build_network}}.}
}
\value{
a list of length \code{npred}. Each element in the list is a dataframe containing
genes' community belongings.
}
\description{
\code{gene_community_detection()} detects genes' community belongings
at each predicted time point.
}
