% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_expression.R
\name{gene_expression}
\alias{gene_expression}
\title{Get individual gene expression across cells}
\usage{
gene_expression(trajectory_expression)
}
\arguments{
\item{trajectory_expression}{a list of length \eqn{N},where \eqn{N} is
the number of cells in one time point(the number of cells need to be equal
among different time points). The \eqn{i}-th element in the list is a matrix with
\eqn{d} rows and \eqn{t} columns, giving the gene expression of \eqn{d} genes in \eqn{t} time
points of the \eqn{i}-th cell lineage trajectory. This list
could be generated by function \code{\link{trajectory_expression}}.}
}
\value{
a list of length \eqn{d}, where \eqn{d} is the number of genes. The elements
in the list representing the individual gene expression in \eqn{N} cell
lineages trajectories, each trajectory has \eqn{t} time points.
}
\description{
\code{gene_expression()} returns individual gene expression across
cell lineages trajectories.
}
