% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_linked_traj.R
\name{plot_linked_traj}
\alias{plot_linked_traj}
\title{Plot Umap with linked trajectories}
\usage{
plot_linked_traj(
  umap_expression,
  cell_name_traj,
  t0 = NULL,
  ntraj = 10,
  random_seed = 2022
)
}
\arguments{
\item{umap_expression}{a dataframe including four columns. The first column is the cell name
consistent with the name in \code{cell_name_traj}. The second and third columns are gene expressions which
dimension is reduced to two dimensions by Umap. The forth column is integers start from 0,
indicating which time points the cell belongs to. The name of four columns need to be
'id','Umap1','Umap2','time',respectively.}

\item{cell_name_traj}{a dataframe with \eqn{N} rows and \eqn{t} columns, giving \eqn{N} unique
trajectories for individual cells, each trajectory has length \eqn{t}. This dataframe could be
generated by function \code{\link{cell_name_traj}}.}

\item{t0}{\code{plot_linked_traj} plots trajectories from time 0 to time \eqn{t_0}. Default setting is
plotting the whole trajectories, which is from time 0 to time \eqn{t-1}.}

\item{ntraj}{an integer indicating number of trajectories to be plotted. Trajectories are randomly
selected from dataframe \code{cell_name_traj}.}

\item{random_seed}{random seed for sampling the trajectories from dataframe \code{cell_name_traj}.}
}
\value{
a plot containing linked trajectories on two dimensional Umap data.
}
\description{
\code{plot_linked_traj()} plots linked trajectories on gene expression data, which
dimension is reduced to two dimensions by Umap.
}
