% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_cell_expression.R
\name{smooth_cell_expression}
\alias{smooth_cell_expression}
\title{Smoothing gene expression patterns for individual cells}
\usage{
smooth_cell_expression(trajectory_expression, npred = 300)
}
\arguments{
\item{trajectory_expression}{a list of length \eqn{N}. The \eqn{i}-th element in the list is a matrix with
\eqn{d} rows and \eqn{t} columns, giving the gene expression of \eqn{d} genes in \eqn{t} time
points of the \eqn{i}-th cell lineage trajectory. This list
could be generated by function \code{\link{trajectory_expression}}.}

\item{npred}{an integer representing number of predicting points equally spaced between
1 and \eqn{t}.}
}
\value{
a list of length two. The first element of the return contains the
predicted values of \code{npred} points in each cell trajectory. The second element of the
return contains the fitted values of \eqn{t} time points in each cell trajectory.
}
\description{
\code{smooth_cell_expression()} returns smoothing gene expression patterns for
individual cells by smoothing spline models. SSANOVA is used in this package.
Both fitted values at \eqn{t} time points and predicted values at
\code{npred} points are returned. \code{smooth_cell_expression()} first smooths the
gene expressions pattern for each gene along time, and extract the
”mean curve” of all individual gene expression patterns to obtain the general
gene expression pattern, which is the output of \code{smooth_cell_expression()}.
}
