% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory_expression.R
\name{trajectory_expression}
\alias{trajectory_expression}
\title{Gene expression in each cell lineage trajectory}
\usage{
trajectory_expression(data, cell_name_traj)
}
\arguments{
\item{data}{a dataframe with \eqn{N*t} rows and \eqn{d+2} columns, where \eqn{N} is
the number of cells in one time point(the number of cells need to be equal
among different time points),
\eqn{t} is the number of time points, \eqn{d} is the number of genes. The first column
is the cell names and the last column is the time points which contains continuous
integers beginning from 1. The column names of the first and the last columns need to be set as
'index' and 'time', respectively.
This dataframe could be regraded as a result of \code{rbind()} of cell-gene
dataset ordered by \emph{increasing} time points.}

\item{cell_name_traj}{a dataframe with \eqn{N} rows and \eqn{t} columns, giving \eqn{N} unique
trajectories for individual cells, each trajectory has length \eqn{t}. This dataframe could be
generated by function \code{\link{cell_name_traj}}.}
}
\value{
a list of length \eqn{N}. The \eqn{i}-th element in the list is a matrix with
\eqn{d} rows and \eqn{t} columns, giving the gene expression of \eqn{d} genes in \eqn{t} time
points of the \eqn{i}-th cell lineage trajectory.
}
\description{
\code{trajectory_expression()} returns a list containing gene expression
of each cell lineage trajectory.
}
